AI-Chemist/requirements.txt
streamlit
python-dotenv
requests
pillow
